This commit will break your project (don't freak out, see desc.)
So I finally got the build.xml file to automatically update with the
library! Just build the project to grab the latest copy of the robo
lib! Only catch is since I don’t know how your file structures look…
you’re going to have to put in the path to the library by hand. It
should be pretty obvious what you have to do, but if it isn’t just ask.
As a note, I have no idea if this works on Windows or not, chances are
no. If @mananshah99 or @Arcterus would be so kind as to check and/or
provide a fix that would be awesome. By the way this means that you
must have the library downloaded for robot code to work… just saying.